+++ src/ReflexState.sol
@@ -17,8 +17,8 @@ import {ReflexConstants} from "./ReflexConstants.sol";
  * | _reentrancyLock | uint256                                               | 0    | 0      | 32    |
  * | _owner          | address                                               | 1    | 0      | 20    |
  * | _pendingOwner   | address                                               | 2    | 0      | 20    |
- * | _modules        | mapping(uint32 => address)                            | 3    | 0      | 32    |
- * | _proxies        | mapping(uint32 => address)                            | 4    | 0      | 32    |
+ * | _modules        | mapping(uint256 => address)                            | 3    | 0      | 32    |
+ * | _proxies        | mapping(uint256 => address)                            | 4    | 0      | 32    |
  * | _relations      | mapping(address => struct TReflexState.TrustRelation) | 5    | 0      | 32    |
  * | __gap           | uint256[44]                                           | 6    | 0      | 1408  |
  */
@@ -50,14 +50,14 @@ abstract contract ReflexState is IReflexState, ReflexConstants {
      * @dev Module id => module implementation.
      * @dev Slot 3 (32 bytes).
      */
-    mapping(uint32 => address) internal _modules;
+    mapping(uint256 => address) internal _modules;
 
     /**
      * @notice Internal proxy mapping.
      * @dev Module id => proxy address (only for single-proxy modules).
      * @dev Slot 4 (32 bytes).
      */
-    mapping(uint32 => address) internal _proxies;
+    mapping(uint256 => address) internal _proxies;
 
     /**
      * @notice Internal proxy to module relation mapping.
