+++ test/mocks/MockReflexBase.sol
@@ -75,11 +75,11 @@ contract MockReflexBase is ReflexBase {
         require(getReentrancyStatus() == _REENTRANCY_LOCK_UNLOCKED);
     }
 
-    function createProxy(uint32 moduleId_, uint16 moduleType_) public returns (address) {
+    function createProxy(uint256 moduleId_, uint256 moduleType_) public returns (address) {
         return _createProxy(moduleId_, moduleType_);
     }
 
-    function callInternalModule(uint32 moduleId_, bytes memory input_) public returns (bytes memory) {
+    function callInternalModule(uint256 moduleId_, bytes memory input_) public returns (bytes memory) {
         return _callInternalModule(moduleId_, input_);
     }
